Configure only OpenTelemetry components
Had to skip the sampler as it depends on the Sentry client. Noted that
after configuring this, Sentry errors now include the root span's name
rather than the name of the source file the exception was thrown in.

The following errors are still reported by the Sentry SDK:

[0] Sentry Logger [error]: You have to set up the SentryContextManager. Without this, the OpenTelemetry & Sentry integration will not work properly.
[0] Sentry Logger [error]: You have to set up the SentryPropagator. Without this, the OpenTelemetry & Sentry integration will not work properly.
[0] Sentry Logger [warn]: You have to set up the SentrySampler. Without this, the OpenTelemetry & Sentry integration may still work, but sample rates set for the Sentry SDK will not be respected. If you use a custom sampler, make sure to use `wrapSamplingDecision`

Only the SentrySampler one is expected.